requirements.txt
fastapi==0.135.1
uvicorn==0.42.0
psycopg2-binary==2.9.11
httpx==0.28.1
uuid6==2024.7.10
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.2.2
python-multipart==0.0.22
